rename test-runner folders
