+++ test-runner/lib/filters/es6_promises_supported_filter.js
