--- test-runner/metamocha/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-db/
-.DS_Store
