--- test-runner/metamocha/index.js
@@ -1,4 +1,0 @@
-'use strict';
-
-exports.Metamocha = require('./lib/metamocha');
-exports.MetadataUi = require('./lib/metadata_ui');
