--- test-runner/metamocha/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "metamocha",
-  "version": "1.3.2",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "npm run lint && node test/test_runner.js",
-    "lint": "eslint lib test"
-  },
-  "author": "Marcus Todd",
-  "license": "MIT",
-  "dependencies": {
-    "mocha": "^3.5.3"
-  },
-  "devDependencies": {
-    "chai": "^4.1.2",
-    "eslint": "^4.7.0"
-  }
-}
