--- test-runner/mongodb-test-runner/README.md
@@ -1,1 +1,0 @@
-# mongodb-test-runner
