--- test-runner/mongodb-test-runner/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-  ConfigurationBase: require('./lib/configuration_base'),
-  EnvironmentBase: require('./lib/environment_base'),
-  topologyManagers: require('mongodb-topology-manager')
-};
