--- test-runner/mongodb-test-runner/lib/filters/travis_filter.js
@@ -1,27 +1,0 @@
-'use strict';
-
-/**
- * Filter for whether Travis should run the test
- *
- * example:
- * metadata: {
- *    ignore: {
- *      travis: true | false
- *    }
- * }
- */
-class TravisFilter {
-  constructor(name) {
-    this.name = name || 'TRAVIS_JOB_ID';
-  }
-
-  filter(test) {
-    if (!test.metadata) return true;
-    if (!test.metadata.ignore) return true;
-    if (!test.metadata.ignore.travis) return true;
-    if (process.env[this.name] !== null && test.metadata.ignore.travis === true) return false;
-    return true;
-  }
-}
-
-module.exports = TravisFilter;
