--- test-runner/mongodb-test-runner/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "mongodb-test-runner",
-  "version": "1.3.4",
-  "description": "",
-  "main": "index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mongodb-js/mongodb-test-runner.git"
-  },
-  "bin": {
-    "mongodb-test-runner": "./lib/runner.js"
-  },
-  "author": "Marcus Todd",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mongodb-js/mongodb-test-runner/issues"
-  },
-  "scripts": {
-    "format": "prettier --print-width 100 --tab-width 2 --single-quote --write lib/**/*.js lib/*.js",
-    "lint": "eslint lib"
-  },
-  "homepage": "https://github.com/mongodb-js/mongodb-test-runner#readme",
-  "devDependencies": {
-    "eslint": "^4.3.0",
-    "eslint-plugin-prettier": "^2.2.0",
-    "prettier": "^1.6.1"
-  },
-  "dependencies": {
-    "metamocha": "^1.2.5",
-    "mongodb-topology-manager": "^2.0.0",
-    "mongodb-version-manager": "^1.0.7",
-    "semver": "^5.4.1",
-    "yargs": "^8.0.2"
-  },
-  "peerOptionalDependencies": {
-    "kerberos": "~0.0",
-    "snappy": "^6.0.0",
-    "bson-ext": "1.0.5"
-  }
-}
