+++ test-runner/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test-runner-poc",
+  "name": "test-runner",
   "version": "1.0.0",
   "description": "",
   "main": "runner.js",
