+++ test-runner/test/crud_tests.js
@@ -1,4 +1,6 @@
 'use strict';
+
+const MongoClient = require('../../lib/mongo_client');
 /* include stuff */
 
 describe('CRUD', function() {
@@ -6,10 +8,10 @@ describe('CRUD', function() {
   it('should insert a document', {
     metadata: { requires: { mongodb: '>=3.6.0', topology: 'single' } },
     test: function() {
-      // const client = new MongoClient('mongodb://localhost:27017');
-      // return client.connect()
-      //   .then(() => insert)
-      //   .then(() => find);
+      const client = new MongoClient('mongodb://localhost:27017');
+      return client.connect()
+        .then(() => insert)
+        .then(() => find);
     }
   });
 });
