+++ test-runner/test/metadata_tests.js
