+++ test/functional/sessions_tests.js
@@ -2,9 +2,9 @@
 const expect = require('chai').expect;
 const setupDatabase = require('./shared').setupDatabase;
 const path = require('path');
-const TestRunnerContext = require('./runner').TestRunnerContext;
-const gatherTestSuites = require('./runner').gatherTestSuites;
-const generateTopologyTests = require('./runner').generateTopologyTests;
+const TestRunnerContext = require('./spec-runner').TestRunnerContext;
+const gatherTestSuites = require('./spec-runner').gatherTestSuites;
+const generateTopologyTests = require('./spec-runner').generateTopologyTests;
 
 const ignoredCommands = ['ismaster'];
 const test = { commands: { started: [], succeeded: [] } };
